pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Assignment"
authors = ["jeha00 <rudtls0611@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.2"
django = "^4.2.7"
mysqlclient = "2.1.1"
mysql-connector-python = "^8.2.0"
drf-spectacular = "^0.26.5"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.develop.dependencies]
add-trailing-comma = "^3.1.0"
black = "^23.11.0"
pre-commit = "^3.5.0"
django-debug-toolbar = "^4.2.0"


[tool.poetry.group.test.dependencies]
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
